old/requirements.txt
reportlab>=4.0.0
python-pptx>=0.6.21